// lionpay-management-old/src/generated-api/wallet/api.ts
export interface CreateMerchantRequest {
    'merchantName': string;
    'countryCode': string;
    'merchantCategory': string;
}
// ...
export interface MerchantModel {
    'merchantId'?: string;
    'merchantName'?: string;
    'countryCode'?: string;
    'merchantCategory'?: string;
    'merchantStatus'?: string;
    'createdAt'?: string;
}
export interface MerchantResponse {
    'merchantId': string;
    'merchantName': string;
    'countryCode': string;
    'merchantCategory': string;
    'merchantStatus': string;
    'createdAt': string;
}
// ...
export interface UpdateMerchantRequest {
    'merchantName': string;
    'merchantCategory': string;
    'merchantStatus': string;
}